Minor changes to get rid of VS2015 warning
BugId:  b2/25824455
Tested: windows 8/Mac OS / Android

+++ component_library/src/rendermesh.cpp
@@ -238,7 +238,7 @@ void RenderMeshComponent::AddFromRawData(corgi::EntityRef& entity,
   // Allocate the array to hold shader default pose's transforms.
   assert(rendermesh_data->shader_transforms == nullptr);
   const uint8_t num_shader_transforms =
-      rendermesh_data->mesh->num_shader_bones();
+      static_cast<uint8_t>(rendermesh_data->mesh->num_shader_bones());
   rendermesh_data->num_shader_transforms = num_shader_transforms;
   if (num_shader_transforms > 0) {
     rendermesh_data->shader_transforms =
